+++ src/timer.c
@@ -5,8 +5,8 @@
 #include "uart1.h"
 #include "utli.h"
 
-extern void set_core_timer_int(unsigned long long s);
-extern void set_core_timer_int_sec(unsigned int s);
+extern void set_core_timer_int(uint64_t s);
+extern void set_core_timer_int_sec(uint32_t s);
 extern void core0_timer_interrupt_enable();
 extern void core0_timer_interrupt_disable();
 
@@ -19,7 +19,7 @@ void print_message(char* msg) {
   print_timestamp();
 }
 
-void add_timer(timer_callback cb, char* msg, unsigned int sec) {
+void add_timer(timer_callback cb, char* msg, uint32_t sec) {
   timer_event* new_event = (timer_event*)simple_malloc(sizeof(timer_event));
 
   if (!new_event) {
